Refactor mostly done. EmailSender class created.
