+++ authorization/src/main/java/com/terabite/authorization/controller/AuthorizationController.java
@@ -4,7 +4,11 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.terabite.authorization.Payload;
 import com.terabite.authorization.repository.UserRepository;
+import com.terabite.authorization.service.EmailSender;
 import com.terabite.authorization.service.UserInformation;
+
+import jakarta.servlet.http.HttpServletRequest;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -16,6 +20,9 @@ import org.springframework.web.service.annotation.HttpExchange;
 public class AuthorizationController {
     @Autowired
     private UserRepository userRepository;
+
+    @Autowired
+    private EmailSender emailSender;
 
     @PostMapping("/signup")
     @ResponseStatus(HttpStatus.CREATED)
@@ -33,5 +40,15 @@ public class AuthorizationController {
     @PostMapping("/logout")
     public Payload userLogoutPost() {
         return new Payload("Reached logout POST");
+    }
+
+    @PostMapping("/forgot_password")
+    public void processForgotPassword(HttpServletRequest request){
+        emailSender.sendForgotPasswordEmail(request);
+    }
+
+    @PostMapping("/reset_password")
+    public void resetPassword(){
+        
     }
 }
