+++ authorization/src/main/java/com/terabite/authorization/repository/LoginRepository.java
@@ -3,9 +3,10 @@ package com.terabite.authorization.repository;
 import com.terabite.authorization.model.Login;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.Optional;
 
 public interface LoginRepository extends JpaRepository<Login, Long> {
     Login findByEmail(String email);
+
+    Login findByPasswordResetToken(String passwordResetToken);
 
 }
